Fix settings window
*Correct closing
*Hierarchy parent-child
